refatoração das funções de filtrar
